Chord detection

+++ inc/helper/Filter.h
@@ -342,11 +342,11 @@ private:
     {
         T* pfCoeff = 0;     // binomial coefficients
 
-        CVector::alloc<T>(pfCoeff, 2 * iOrder);
+        CVector::alloc<T>(pfCoeff, static_cast<long long>(2) * iOrder);
 
         for (auto j = 0; j < iOrder; j++)
         {
-            T fArg = static_cast<T>(M_PI * (2 * j + 1) / (2 * iOrder));
+            T fArg = static_cast<T>(M_PI * (2 * j + 1) / (2. * iOrder));
             T fNorm = static_cast<T>(1 + std::sin(M_PI * fCutOff) * std::sin(fArg));
             pfCoeff[2 * j] = static_cast<T>(-std::cos(M_PI * fCutOff) / fNorm);
             pfCoeff[2 * j + 1] = static_cast<T>(-std::sin(M_PI * fCutOff) * std::cos(fArg) / fNorm);
